fix getter bug.

+++ ts-train-service/src/main/java/train/domain/Information.java
@@ -31,7 +31,7 @@ public class Information {
         this.id = id;
     }
 
-    public long getEconomyClass() {
+    public int getEconomyClass() {
         return economyClass;
     }
